all methods dispositivo

+++ src/main/java/epicode/gestioneDispositivi/repositories/DispositivoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DispositivoRepository extends JpaRepository<Dispositivo,Integer > {
+public interface DispositivoRepository extends JpaRepository<Dispositivo,Long > {
 
 
 }
